% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredDrugInducedKidneyInjury.R
\docType{package}
\name{PredDrugInducedKidneyInjury}
\alias{PredDrugInducedKidneyInjury}
\title{PredDrugInducedKidneyInjury}
\description{
A package for running prediction model development studies using data in the OMOP CDM
}
